fix(store): throw on compaction-busy snapshot save; validate data/ on load (#3162) - Addressed review comments. Added few Unit test cases, -Removed test-snapshot-corruption.sh because scenario is already covered by UTs.
