+++ hugegraph-store/hg-store-test/src/main/java/org/apache/hugegraph/store/core/snapshot/HgSnapshotHandlerTest.java
@@ -18,11 +18,7 @@
 package org.apache.hugegraph.store.core.snapshot;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertThrows;
-import static org.junit.Assert.assertTrue;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 
 import java.io.File;
 import java.io.IOException;
@@ -31,17 +27,18 @@ import java.nio.file.Files;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.atomic.AtomicInteger;
 
 import org.apache.commons.io.FileUtils;
-import org.apache.hugegraph.store.HgStoreEngine;
-import org.apache.hugegraph.store.PartitionEngine;
-import org.apache.hugegraph.store.business.BusinessHandler;
 import org.apache.hugegraph.store.core.StoreEngineTestBase;
 import org.apache.hugegraph.store.meta.Partition;
 import org.apache.hugegraph.store.snapshot.HgSnapshotHandler;
 import org.apache.hugegraph.store.snapshot.SnapshotHandler;
 import org.apache.hugegraph.store.util.HgStoreException;
+
+import com.alipay.sofa.jraft.entity.RaftOutter;
+import com.alipay.sofa.jraft.storage.snapshot.SnapshotReader;
+import com.alipay.sofa.jraft.storage.snapshot.SnapshotWriter;
+import com.google.protobuf.Message;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -65,152 +62,6 @@ public class HgSnapshotHandlerTest extends StoreEngineTestBase {
         hgSnapshotHandlerUnderTest = new HgSnapshotHandler(createPartitionEngine(0));
         FileUtils.forceMkdir(new File("/tmp/snapshot"));
         FileUtils.forceMkdir(new File("/tmp/snapshot/data"));
-    }
-
-    // ── Fix 1: onSnapshotSave must throw when compaction is in progress ────────
-
-    /**
-     * Before the fix, onSnapshotSave silently returned when state == doing,
-     * causing JRaft to commit an empty snapshot dir with no data/.
-     * After the fix it must throw HgStoreException so JRaft retries instead.
-     */
-    @Test
-    public void testOnSnapshotSaveThrowsWhenCompactionInProgress() {
-        // Build a SnapshotHandler wired to a mock PartitionEngine whose BusinessHandler
-        // reports state == doing (compaction active) for partition 0.
-        PartitionEngine mockEngine = mock(PartitionEngine.class);
-        HgStoreEngine mockStoreEngine = mock(HgStoreEngine.class);
-        BusinessHandler mockBusinessHandler = mock(BusinessHandler.class);
-
-        AtomicInteger doingState = new AtomicInteger(BusinessHandler.doing);
-
-        when(mockEngine.getGroupId()).thenReturn(0);
-        when(mockEngine.getStoreEngine()).thenReturn(mockStoreEngine);
-        when(mockStoreEngine.getBusinessHandler()).thenReturn(mockBusinessHandler);
-        when(mockBusinessHandler.getState(0)).thenReturn(doingState);
-
-        SnapshotHandler handler = new SnapshotHandler(mockEngine);
-
-        SnapshotWriter stubWriter = stubWriter("/tmp/snapshot");
-
-        HgStoreException ex = assertThrows(
-                "onSnapshotSave must throw when state == doing",
-                HgStoreException.class,
-                () -> handler.onSnapshotSave(stubWriter));
-
-        assertTrue("Exception message must mention the partition",
-                   ex.getMessage().contains("0"));
-        assertTrue("Exception message must describe the cause",
-                   ex.getMessage().contains("compaction in progress"));
-    }
-
-    /**
-     * When state is NOT doing (e.g. compactionDone), onSnapshotSave must not throw.
-     */
-    @Test
-    public void testOnSnapshotSaveDoesNotThrowWhenNotBusy() throws Exception {
-        PartitionEngine mockEngine = mock(PartitionEngine.class);
-        HgStoreEngine mockStoreEngine = mock(HgStoreEngine.class);
-        BusinessHandler mockBusinessHandler = mock(BusinessHandler.class);
-
-        // state == compactionDone (not doing) — save should proceed normally
-        AtomicInteger doneState = new AtomicInteger(BusinessHandler.compactionDone);
-
-        when(mockEngine.getGroupId()).thenReturn(0);
-        when(mockEngine.getStoreEngine()).thenReturn(mockStoreEngine);
-        when(mockStoreEngine.getBusinessHandler()).thenReturn(mockBusinessHandler);
-        when(mockBusinessHandler.getState(0)).thenReturn(doneState);
-
-        // saveSnapshot is a no-op via the mock, so we just need it not to throw at the guard
-        SnapshotHandler handler = new SnapshotHandler(mockEngine);
-        SnapshotWriter stubWriter = stubWriter(tmpDir.newFolder("snap-not-busy").getAbsolutePath());
-
-        // No exception should propagate from the state guard.
-        // (saveSnapshot will throw because the mock returns null for it — that's fine,
-        //  we only care the doing-check is not hit.)
-        try {
-            handler.onSnapshotSave(stubWriter);
-        } catch (HgStoreException e) {
-            assertFalse("Must not be the compaction-busy exception",
-                        e.getMessage().contains("compaction in progress"));
-        }
-    }
-
-    // ── Fix 2: onSnapshotLoad must not silently skip a corrupt snapshot ────────
-
-    /**
-     * Before the fix, onSnapshotLoad returned silently when should_not_load was present,
-     * even if data/ was missing — leaving the partition in an undefined state.
-     * After the fix it must fall through to the real load path and throw,
-     * so JRaft can signal the error and the leader can install a fresh snapshot.
-     */
-    @Test
-    public void testOnSnapshotLoadFallsThroughWhenShouldNotLoadPresentButDataMissing()
-            throws Exception {
-        // Arrange: snapshot dir has should_not_load but NO data/ subdirectory.
-        File snapDir = tmpDir.newFolder("snapshot-corrupt");
-        File shouldNotLoad = new File(snapDir, "should_not_load");
-        Files.write(shouldNotLoad.toPath(), "saved snapshot".getBytes(StandardCharsets.UTF_8));
-        // data/ deliberately not created
-
-        SnapshotHandler handler = new SnapshotHandler(createPartitionEngine(1));
-        SnapshotReader stubReader = stubReader(snapDir.getAbsolutePath());
-
-        // The fix causes execution to fall through shouldNotLoad() and call
-        // businessHandler.loadSnapshot(missingDataDir) which throws HgStoreException.
-        assertThrows(
-                "onSnapshotLoad must throw when should_not_load present but data/ missing",
-                HgStoreException.class,
-                () -> handler.onSnapshotLoad(stubReader, 0L));
-    }
-
-    /**
-     * When should_not_load is present AND data/ also exists, onSnapshotLoad must
-     * return early (normal locally-saved snapshot — no load needed).
-     */
-    @Test
-    public void testOnSnapshotLoadSkipsWhenShouldNotLoadPresentAndDataExists() throws Exception {
-        // Arrange: a healthy local snapshot — both should_not_load and data/ present.
-        File snapDir = tmpDir.newFolder("snapshot-healthy");
-        File shouldNotLoad = new File(snapDir, "should_not_load");
-        Files.write(shouldNotLoad.toPath(), "saved snapshot".getBytes(StandardCharsets.UTF_8));
-        FileUtils.forceMkdir(new File(snapDir, "data"));
-
-        SnapshotHandler handler = new SnapshotHandler(createPartitionEngine(2));
-        SnapshotReader stubReader = stubReader(snapDir.getAbsolutePath());
-
-        // Must not throw — should return early at the should_not_load + data-exists check.
-        handler.onSnapshotLoad(stubReader, 0L);
-    }
-
-    // ── Stub helpers ──────────────────────────────────────────────────────────
-
-    private static SnapshotWriter stubWriter(String path) {
-        return new SnapshotWriter() {
-            @Override public boolean saveMeta(RaftOutter.SnapshotMeta meta) { return false; }
-            @Override public boolean addFile(String fileName, Message fileMeta) { return false; }
-            @Override public boolean removeFile(String fileName) { return false; }
-            @Override public void close(boolean keepDataOnError) {}
-            @Override public boolean init(Void opts) { return false; }
-            @Override public void shutdown() {}
-            @Override public String getPath() { return path; }
-            @Override public Set<String> listFiles() { return null; }
-            @Override public Message getFileMeta(String fileName) { return null; }
-            @Override public void close() {}
-        };
-    }
-
-    private static SnapshotReader stubReader(String path) {
-        return new SnapshotReader() {
-            @Override public RaftOutter.SnapshotMeta load() { return null; }
-            @Override public String generateURIForCopy() { return null; }
-            @Override public boolean init(Void opts) { return false; }
-            @Override public void shutdown() {}
-            @Override public String getPath() { return path; }
-            @Override public Set<String> listFiles() { return null; }
-            @Override public Message getFileMeta(String fileName) { return null; }
-            @Override public void close() {}
-        };
     }
 
     @Test
@@ -347,5 +198,59 @@ public class HgSnapshotHandlerTest extends StoreEngineTestBase {
         HgSnapshotHandler.findFileList(dir, rootDir, List.of("value"));
 
         // Verify the results
+    }
+
+    /**
+     * Test that onSnapshotLoad validates corruption when should_not_load is present but data/ missing.
+     */
+    @Test
+    public void testOnSnapshotLoadFallsThroughWhenShouldNotLoadPresentButDataMissing()
+            throws Exception {
+        // Arrange: snapshot dir has should_not_load but NO data/ subdirectory.
+        File snapDir = tmpDir.newFolder("snapshot-corrupt");
+        File shouldNotLoad = new File(snapDir, "should_not_load");
+        Files.write(shouldNotLoad.toPath(), "saved snapshot".getBytes(StandardCharsets.UTF_8));
+        // data/ deliberately not created
+
+        SnapshotHandler handler = new SnapshotHandler(createPartitionEngine(1));
+        SnapshotReader stubReader = stubReader(snapDir.getAbsolutePath());
+
+        // The fix causes execution to fall through shouldNotLoad() and call
+        // businessHandler.loadSnapshot(missingDataDir) which throws HgStoreException.
+        assertThrows(
+                "onSnapshotLoad must throw when should_not_load present but data/ missing",
+                HgStoreException.class,
+                () -> handler.onSnapshotLoad(stubReader, 0L));
+    }
+
+    /**
+     * Test that onSnapshotLoad skips loading when snapshot is locally saved (both flags present).
+     */
+    @Test
+    public void testOnSnapshotLoadSkipsWhenShouldNotLoadPresentAndDataExists() throws Exception {
+        // Arrange: a healthy local snapshot — both should_not_load and data/ present.
+        File snapDir = tmpDir.newFolder("snapshot-healthy");
+        File shouldNotLoad = new File(snapDir, "should_not_load");
+        Files.write(shouldNotLoad.toPath(), "saved snapshot".getBytes(StandardCharsets.UTF_8));
+        FileUtils.forceMkdir(new File(snapDir, "data"));
+
+        SnapshotHandler handler = new SnapshotHandler(createPartitionEngine(2));
+        SnapshotReader stubReader = stubReader(snapDir.getAbsolutePath());
+
+        // Must not throw — should return early at the should_not_load + data-exists check.
+        handler.onSnapshotLoad(stubReader, 0L);
+    }
+
+    private static SnapshotReader stubReader(String path) {
+        return new SnapshotReader() {
+            @Override public RaftOutter.SnapshotMeta load() { return null; }
+            @Override public String generateURIForCopy() { return null; }
+            @Override public boolean init(Void opts) { return false; }
+            @Override public void shutdown() {}
+            @Override public String getPath() { return path; }
+            @Override public Set<String> listFiles() { return null; }
+            @Override public Message getFileMeta(String fileName) { return null; }
+            @Override public void close() {}
+        };
     }
 }
